% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltim.R
\docType{data}
\name{baltim}
\alias{baltim}
\title{Baltimore house sales prices and hedonics}
\format{
A data frame with 211 rows and 17 variables:
\describe{
  \item{STATION}{ID variable}
  \item{PRICE}{sales price, in 1,000 US dollars (MLS)}
  \item{NROOM}{number of rooms}
  \item{DWELL}{1 if detached unit, 0 otherwise}
  \item{NBATH}{number of bathrooms}
  \item{PATIO}{1 if patio, 0 otherwise}
  \item{FIREPL}{1 if fireplace, 0 otherwise}
  \item{AC}{1 if air conditioning, 0 otherwise}
  \item{BMENT}{1 if basement, 0 otherwise}
  \item{NSTOR}{number of stores}
  \item{GAR}{number of car space in garage, (0 = no garage)}
  \item{AGE}{age of dwellings in years}
  \item{CITCOU}{1 if dwelling is in Baltimore County, 0 otherwise}
  \item{LOTSZ}{lot size in hundreds of square feet}
  \item{SQFT}{interior living space in hundreds of square feet}
  \item{X}{X coordinate on the Maryland grid}
  \item{Y}{Y coordinate on the Maryland grid}
}
}
\source{
\url{https://geodacenter.github.io/data-and-lab/}
}
\usage{
baltim
}
\description{
A dataset containing the prices and other attributes of 211 dwelling in Baltimore, MD
}
\keyword{datasets}
