% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_regimes.R
\name{error_regimes}
\alias{error_regimes}
\alias{coef.error_regimes}
\alias{vcov.error_regimes}
\alias{print.error_regimes}
\alias{summary.error_regimes}
\alias{print.summary.error_regimes}
\title{Estimation of spatial regime models}
\usage{
error_regimes(
  formula,
  data,
  listw,
  rgv,
  weps_rg = FALSE,
  initial.value = NULL,
  het,
  verbose = FALSE,
  control,
  cl
)

\method{coef}{error_regimes}(object, ...)

\method{vcov}{error_regimes}(object, ...)

\method{print}{error_regimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{error_regimes}(object, ...)

\method{print}{summary.error_regimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix}

\item{rgv}{variable to identify the regimes}

\item{weps_rg}{default FALSE, if TRUE the spatial error term varies by regimes (see details)}

\item{initial.value}{starting point for the optimization}

\item{het}{heteroskedastic variance-covariance matrix}

\item{verbose}{print a trace of the optimization}

\item{control}{argument for optimization}

\item{cl}{record calls}

\item{object}{an object of class lag_regime}

\item{...}{additional arguments}

\item{x}{an object of class lag_regime}

\item{digits}{number of digits}
}
\description{
Estimation of spatial regime models
}
