% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsar2sls.R
\name{hsar2sls}
\alias{hsar2sls}
\alias{summary.hsar2sls}
\alias{print.summary.hsar2sls}
\title{Estimation of HSAR models by 2SLS}
\usage{
hsar2sls(formula, data, listw = NULL, index = NULL, nins = 2, ...)

\method{summary}{hsar2sls}(object, MG = TRUE, ...)

\method{print}{summary.hsar2sls}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{pdata.frame}.}

\item{listw}{object. An object of class \code{listw}, \code{matrix}, or \code{Matrix}.}

\item{index}{index.}

\item{nins}{numeric. Number of instrument. \code{nins = 2} as default.}

\item{...}{additional arguments passed to \code{maxLik}}

\item{MG}{logical. If \code{TRUE}, the Mean Group estimator is returned}

\item{x, object}{an object of class \code{hsar2sls}}

\item{digits}{the number of digits}
}
\description{
Estimation of HSAR models by 2SLS
}
