% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsarML.R
\name{hsarML}
\alias{hsarML}
\alias{coef.hsarML}
\alias{summary.hsarML}
\alias{print.summary.hsarML}
\title{Estimation of HSAR models by Quasi-Maximum Likelihood}
\usage{
hsarML(
  formula,
  data,
  listw = NULL,
  index = NULL,
  gradient = TRUE,
  average = FALSE,
  init.values = NULL,
  print.init = FALSE,
  otype = c("maxLik", "optim"),
  ...
)

\method{coef}{hsarML}(object, ...)

\method{summary}{hsarML}(object, MG = TRUE, ...)

\method{print}{summary.hsarML}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{pdata.frame}.}

\item{listw}{object. An object of class \code{listw}, \code{matrix}, or \code{Matrix}.}

\item{index}{index.}

\item{gradient}{logical. Only for testing procedures. Should the analytic gradient be used in the ML optimization procedure? \code{TRUE} as default. If \code{FALSE}, then the numerical gradient is used.}

\item{average}{logical. Should the sample log-likelihood function be divided by N?}

\item{init.values}{if not \code{NULL}, the user must provide a vector of initial parameters for the optimization procedure.}

\item{print.init}{logical. If \code{TRUE} the initial parameters used in the optimization of the first step are printed.}

\item{otype}{string. A string indicating whether package \code{maxLik} or \code{optim} is used in for the numerical optimization.}

\item{...}{additional arguments passed to \code{maxLik}}

\item{MG}{logical. If \code{TRUE}, the Mean Group estimator is returned}

\item{x, object}{an object of class \code{hsarML}}

\item{digits}{the number of digits}
}
\description{
Estimation of HSAR models by Quasi-Maximum Likelihood
}
