% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregimes.R
\name{ivregimes}
\alias{ivregimes}
\alias{coef.ivregimes}
\alias{vcov.ivregimes}
\alias{print.ivregimes}
\alias{summary.ivregimes}
\alias{print.summary.ivregimes}
\title{Estimation of spatial regime models}
\usage{
ivregimes(
  formula,
  data,
  rgv = NULL,
  endog = NULL,
  instr = NULL,
  endog_v = NULL,
  vc = c("classical", "robust", "OGMM")
)

\method{coef}{ivregimes}(object, ...)

\method{vcov}{ivregimes}(object, ...)

\method{print}{ivregimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{ivregimes}(object, ...)

\method{print}{summary.ivregimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{rgv}{variable to identify the regimes}

\item{endog}{formula containing  endogenous variables}

\item{instr}{formula containing instrument for endogenous variables}

\item{endog_v}{logical vector defining which endogenous variable should vary by regime}

\item{vc}{one of ("classical", "robust", "OGMM")}

\item{object}{an object of class ivregime}

\item{...}{additional arguments}

\item{x}{an object of class ivregime}

\item{digits}{number of digits}
}
\value{
An object of class ``\code{spregimes}''
}
\description{
Estimation of spatial regime models
}
\details{
The model estimated is:

\deqn{
y_{ij}= \mathbf{x_{ij,k}}\beta_j + \mathbf{Y_{ij,k}}\gamma_j + \epsilon
}
for i=1,..,n representing the sample observations, and j =1,..., J representing
the  regimes
}
\author{
Gianfranco Piras and Mauricio Sarrias
}
