% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_regimes.R
\name{lag_regimes}
\alias{lag_regimes}
\alias{coef.lag_regimes}
\alias{vcov.lag_regimes}
\alias{print.lag_regimes}
\alias{summary.lag_regimes}
\alias{print.summary.lag_regimes}
\title{Estimation of spatial regime models}
\usage{
lag_regimes(formula, data, listw, rgv, het, cl, wy_rg)

\method{coef}{lag_regimes}(object, ...)

\method{vcov}{lag_regimes}(object, ...)

\method{print}{lag_regimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{lag_regimes}(object, ...)

\method{print}{summary.lag_regimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix}

\item{rgv}{variable to identify the regimes}

\item{het}{heteroskedastic variance-covariance matrix}

\item{cl}{record calls}

\item{wy_rg}{default \code{wy_rg = TRUE}, the lagged dependent variable varies for regime}

\item{object}{an object of class lag_regime}

\item{...}{additional arguments}

\item{x}{an object of class lag_regime}

\item{digits}{number of digits}
}
\description{
Estimation of spatial regime models
}
