% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_6.R
\docType{data}
\name{ws_6}
\alias{ws_6}
\title{US Counties Homicides}
\format{
A spatial weighting matrix of class \code{Matrix}
}
\source{
\url{https://geodacenter.github.io/data-and-lab/}
}
\usage{
ws_6
}
\description{
Continental U.S. counties data for homicides and selected socio-economic characteristics. Data for four decennial census years: 1960, 1970, 1980 and 1990.
}
\keyword{datasets}
