% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_regimes.R
\name{error_regimes}
\alias{error_regimes}
\title{Estimation of spatial regime models}
\usage{
error_regimes(
  formula,
  data,
  listw,
  rgv,
  initial.value,
  het,
  step1.c,
  control,
  cl
)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix}

\item{rgv}{variable to identify the regimes}

\item{het}{heteroskedastic variance-covariance matrix}

\item{cl}{record calls}

\item{object}{an object of class lag_regime}

\item{...}{additional arguments}

\item{x}{an object of class lag_regime}

\item{digits}{number of digits}
}
\description{
Estimation of spatial regime models
}
