% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_regimes.R
\name{lag_regimes}
\alias{lag_regimes}
\alias{coef.lag_regimes}
\alias{vcov.lag_regimes}
\alias{print.lag_regimes}
\alias{summary.lag_regimes}
\alias{print.summary.lag_regimes}
\title{Estimation of spatial regimes models}
\usage{
lag_regimes(formula, data, listw, rgv, het, cl, wy_rg)

\method{coef}{lag_regimes}(object, ...)

\method{vcov}{lag_regimes}(object, ...)

\method{print}{lag_regimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{lag_regimes}(object, ...)

\method{print}{summary.lag_regimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model of the form \code{y ~ x_f | x_v | wx | h_f | h_v | wh} where \code{y} is the dependent variable, \code{x_f} are the regressors that do not vary by regimes,  \code{x_v} are the regressors that vary by regimes, \code{wx} are the spatially lagged regressors, \code{h_f} are the instruments that do not vary by regimes,  \code{h_v} are the instruments that vary by regimes, \code{wh} are the spatially lagged instruments.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix of class \code{listw}, \code{matrix} or \code{Matrix}}

\item{rgv}{an object of class \code{formula} to identify the regime variables}

\item{het}{heteroskedastic variance-covariance matrix}

\item{cl}{record calls}

\item{wy_rg}{default \code{wy_rg = FALSE}, the lagged dependent variable does not vary by regime (see details)}

\item{object}{an object of class lag_regime}

\item{...}{additional arguments}

\item{x}{an object of class lag_regime}

\item{digits}{number of digits}
}
\value{
an object of class \code{"lag_regimes"}. A \code{list} with six elements. The first element of the \code{list} contains the estimation results. The other elements are needed for printing.
}
\description{
Estimation of spatial regimes models
}
\examples{
data("natreg")
data("ws_6")
form <-  HR90  ~ 0 | MA90 + PS90 +
RD90 + UE90 | 0 | 0 | MA90 + PS90 +
RD90 + FH90 + FP89 + GI89 | 0

form1 <-  HR90  ~ MA90 -1 |  PS90 +
RD90 + UE90 | 0 | MA90 -1 |  PS90 +
RD90 + FH90 + FP89 + GI89 | 0

split  <- ~ REGIONS

###############################
#  Spatial Lag regimes model  #
###############################
mod4 <- spregimes(formula = form, data = natreg,
rgv = split, listw = ws_6, model = "lag",
het = TRUE, wy_rg = TRUE)
summary(mod4)
mod5 <- spregimes(formula = form1, data = natreg,
rgv = split, listw = ws_6, model = "lag",
het = TRUE, wy_rg = TRUE)
summary(mod5)

}
