% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_regimes.R
\name{ols_regimes}
\alias{ols_regimes}
\alias{coef.ols_regimes}
\alias{vcov.ols_regimes}
\alias{print.ols_regimes}
\alias{summary.ols_regimes}
\alias{print.summary.ols_regimes}
\title{Estimation of spatial regime models}
\usage{
ols_regimes(formula, data, listw, rgv, het, cl)

\method{coef}{ols_regimes}(object, ...)

\method{vcov}{ols_regimes}(object, ...)

\method{print}{ols_regimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{ols_regimes}(object, ...)

\method{print}{summary.ols_regimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix}

\item{rgv}{variable to identify the regimes}

\item{het}{heteroskedastic variance-covariance matrix}

\item{cl}{record calls}

\item{object}{an object of class lag_regime}

\item{...}{additional arguments}

\item{x}{an object of class lag_regime}

\item{digits}{number of digits}
}
\description{
Estimation of spatial regime models
}
