% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes.R
\name{regimes}
\alias{regimes}
\title{Estimation of spatial regime models}
\usage{
regimes(formula, data, rgv = NULL, vc = c("homoskedastic", "groupwise"))
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{rgv}{variable to identify the regimes}

\item{vc}{one of ("homoskedastic", "groupwise")}
}
\value{
An object of class ``\code{lm}'', a list with elements:
}
\description{
Estimation of spatial regime models
}
\details{
The model estimated is:

\deqn{
y_{ij}= \mathbf{x_{ij,k}}\beta_j + \epsilon
}
for i=1,..,n representing the sample observations, and j =1,..., J representing
the  regimes
}
\author{
Gianfranco Piras and Mauricio Sarrias
}
