% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes.R
\name{regimes}
\alias{regimes}
\alias{coef.regimes}
\alias{vcov.regimes}
\alias{print.regimes}
\alias{summary.regimes}
\alias{print.summary.regimes}
\title{Estimation of spatial regime models}
\usage{
regimes(formula, data, rgv = NULL, vc = c("homoskedastic", "groupwise"))

\method{coef}{regimes}(object, ...)

\method{vcov}{regimes}(object, ...)

\method{print}{regimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{regimes}(object, ...)

\method{print}{summary.regimes}(x, digits = max(5, getOption("digits") - 3), ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{rgv}{variable to identify the regimes}

\item{vc}{one of ("homoskedastic", "groupwise")}

\item{object}{an object of class ivregime}

\item{...}{additional arguments}

\item{x}{an object of class ivregime}

\item{digits}{number of digits}
}
\value{
An object of class ``\code{lm}'', a list with elements:
}
\description{
Estimation of spatial regime models
}
\details{
The model estimated is:

\deqn{
y_{ij}= \mathbf{x_{ij,k}}\beta_j + \epsilon
}
for i=1,..,n representing the sample observations, and j =1,..., J representing
the  regimes
}
\examples{
data("baltim")
nbB <- spdep::read.gal(system.file("extdata", "baltimore.gal", package = "spregimes"))
listw1 <- spdep::nb2listw(nbB, style = "W")
form   <- PRICE  ~ NROOM + NBATH + PATIO + FIREPL + AC + GAR + AGE + LOTSZ + SQFT
split  <- ~ CITCOU
mod <- regimes(formula = form, data = baltim, rgv = split, vc = "groupwise")
summary(mod)
form <- PRICE  ~ AC + AGE + NROOM + PATIO + FIREPL + SQFT | NBATH + GAR + LOTSZ - 1
mod <- regimes(form, baltim, split, vc = "groupwise")
summary(mod)


}
\author{
Gianfranco Piras and Mauricio Sarrias
}
