% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes.R
\name{regimes}
\alias{regimes}
\alias{print.spregimes}
\alias{summary.spregimes}
\title{Estimation of spatial regime models}
\usage{
regimes(
  formula,
  data,
  splitv,
  vc = c("homoskedastic", "groupwise", "heteroskedastic")
)

\method{print}{spregimes}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{spregimes}(object, ...)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{splitv}{variable to split the dataset}

\item{vc}{one of ("homoskedastic", "groupwise","heteroskedastic")}

\item{x}{object for printing}

\item{digits}{number of digits}

\item{...}{additional arguments to be passed}

\item{object}{object for printing}
}
\value{
An object of class ``\code{spregimes}'', a list with elements:
\item{coefficients}{the estimated coefficients,}
\item{call}{the matched call,}
\item{X}{the X matrix}
\item{y}{the dependent variable}
#' @import Matrix stats spatialreg methods Formula maxLik sphet
}
\description{
Estimation of spatial regime models
}
\details{
The model estimated is:

\deqn{
y_{ij}= \mathbf{x_{ij,k}}\beta_j + \epsilon
}
for i=1,..,n representing the sample observations, and j =1,..., J representing
the  regimes
}
\author{
Gianfranco Piras and Mauricio Sarrias
}
