% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spregimes.R
\name{spregimes}
\alias{spregimes}
\title{Estimation of spatial regime models}
\usage{
spregimes(
  formula,
  data = list(),
  listw,
  listw2 = NULL,
  rgv = NULL,
  initial.value = 0.2,
  wy_rg = FALSE,
  model = c("sarar", "lag", "error"),
  het = FALSE,
  step1.c = FALSE,
  control = list()
)
}
\arguments{
\item{formula}{a symbolic description of the model.}

\item{data}{the data of class \code{data.frame}.}

\item{listw}{a spatial weighting matrix}

\item{listw2}{an (optional) spatial weighting matrix for sarar models}

\item{rgv}{variable to identify the regimes}

\item{initial.value}{initial value for the spatial error parameter}

\item{wy_rg}{default \code{wy_rg = TRUE}, the lagged dependent variable varies for regime}

\item{model}{one of ("sarar", "lag", "error")}

\item{het}{heteroskedastic variance-covariance matrix}

\item{step1.c}{use with error and sarar}

\item{control}{list of controls for the minimization}
}
\value{
An object of class ``\code{lag_regimes}'', or \code{sarar_regimes}, or \code{error_regimes}
}
\description{
Estimation of spatial regime models
}
\details{
The model estimated is:

\deqn{
y_{ij}= \mathbf{x_{ij,k}}\beta_j + \mathbf{Y_{ij,k}}\gamma_j + \epsilon
}
for i=1,..,n representing the sample observations, and j =1,..., J representing
the  regimes
}
\author{
Gianfranco Piras and Mauricio Sarrias
}
