% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_6.R
\docType{data}
\name{ws_6}
\alias{ws_6}
\title{Spatial weighting matrix for the US Counties Homicides data}
\format{
A spatial weighting matrix of class \code{Matrix}
}
\source{
\url{https://geodacenter.github.io/data-and-lab/}
}
\usage{
ws_6
}
\description{
ws_6 is a spatial weights matrix based on the 6 nearest neighbors for the Continental U.S. counties data for homicides
}
\keyword{datasets}
